Extract TFD validators to traits
