+++ src/CfdiUtils/Validate/Cfdi33/Standard/TimbreFiscalDigitalVersion.php
@@ -2,10 +2,8 @@
 
 namespace CfdiUtils\Validate\Cfdi33\Standard;
 
-use CfdiUtils\Nodes\NodeInterface;
-use CfdiUtils\Validate\Asserts;
 use CfdiUtils\Validate\Cfdi33\Abstracts\AbstractDiscoverableVersion33;
-use CfdiUtils\Validate\Status;
+use CfdiUtils\Validate\Common\TimbreFiscalDigitalVersionValidatorTrait;
 
 /**
  * TimbreFiscalDigitalVersion
@@ -15,19 +13,5 @@ use CfdiUtils\Validate\Status;
  */
 class TimbreFiscalDigitalVersion extends AbstractDiscoverableVersion33
 {
-    public function validate(NodeInterface $comprobante, Asserts $asserts)
-    {
-        $asserts->put(
-            'TFDVERSION01',
-            'Si existe el complemento timbre fiscal digital, entonces su versión debe ser 1.1'
-        );
-
-        $tfdVersion = $comprobante->searchNode('cfdi:Complemento', 'tfd:TimbreFiscalDigital');
-        if (null !== $tfdVersion) {
-            $asserts->putStatus(
-                'TFDVERSION01',
-                Status::when('1.1' === $tfdVersion['Version'])
-            );
-        }
-    }
+    use TimbreFiscalDigitalVersionValidatorTrait;
 }
